refactor: extract `format`
